apitest-python-pytest-requests/requirements.txt
pytest==5.4.3
requests==2.24.0
allure-pytest==2.9.45